move customer and merchant to modules
