+++ frontend/modules/Customer/index.js
@@ -5,8 +5,8 @@ import ProgramIsActive from './ProgramIsActive';
 import ProgramNotActive from './ProgramNotActive';
 import Welcome from './Welcome';
 import { useRouter } from 'next/router';
-import PageBackground from '../PageBackground';
-import Loader from '../Loader';
+import PageBackground from '../../components/PageBackground';
+import Loader from '../../components/Loader';
 import { backend } from '../../utils/backend';
 import { factory } from '../../utils/near-ft-factory';
 import { customer } from '../../utils/customer';
