+++ frontend/modules/Merchant/index.js
@@ -5,8 +5,8 @@ import { getMerchantAddress } from '../../utils/utils';
 import Welcome from './Welcome/Welcome';
 import SignedOut from './SignedOut';
 import SignedIn from './SignedIn';
-import PageBackground from '../PageBackground';
-import Loader from '../Loader';
+import PageBackground from '../../components/PageBackground';
+import Loader from '../../components/Loader';
 import { backend } from '../../utils/backend';
 import { factory } from '../../utils/near-ft-factory';
