+++ frontend/pages/customer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import CustomerView from '../components/Customer';
+import CustomerView from '../modules/Customer';
 
 const Customer = () => <CustomerView />;
